batch_process/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
distlib==0.3.6
distro-info==1.0
ffmpeg==1.4
ffprobe==0.5
filelock==3.8.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-cloud-pubsub==2.13.11
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==4.21.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.4
rabbitmq==0.2.0
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.6
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
